Adapted some pending tests to the new workflow using pathnames
