+++ reflexible/conv2netcdf4/__init__.py
@@ -2,3 +2,4 @@
 from .legacy_structures import Header, Structure, BinaryFile, FDC
 from .flexpart_read import read_header, read_trajectories, read_command
 from .grid_read import read_grid, get_slabs, fill_grids
+from .helpers import get_fpdirs
